Move spruce form into proper directory
